+++ src/server/models/__tests__/test-user.js
@@ -2,7 +2,7 @@ describe('User', () => {
     const {Model} = require('../model');
 
     jest.dontMock('../user');
-    const {User} = require('../user');
+    require('../user');
 
     it('inherits Model', () => {
         expect(Model).toBeCalledWith('users');